+++ src/main/java/com/ohgiraffers/refrigegobackend/bookmark/infrastructure/repository/BookmarkRepository.java
@@ -10,14 +10,14 @@ import java.util.List;
 import java.util.Optional;
 
 public interface BookmarkRepository extends JpaRepository<Bookmark, Long> {
-    Optional<Bookmark> findByUserIdAndRecipeRcpSeq(Integer userId, String rcpSeq);
+    Optional<Bookmark> findByUserIdAndRecipeRcpSeq(Long userId, String rcpSeq);
 
 
     @Query("SELECT l.recipe.rcpSeq FROM Bookmark l WHERE l.user.id = :userId")
-    List<String> findRecipeIdsByUserId(@Param("userId") int userId);
+    List<String> findRecipeIdsByUserId(@Param("userId") Long userId);
 
     @Query("SELECT b.recipe FROM Bookmark b WHERE b.user.id = :userId")
-    List<Recipe> findRecipesByUserId(@Param("userId") int userId);
+    List<Recipe> findRecipesByUserId(@Param("userId") Long userId);
 
-    List<Bookmark> findByUserId(int userId);
+    List<Bookmark> findByUserId(Long userId);
 }
